requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.7
django-appconf==1.0.6
django-fsm==2.8.1
django-fsm-log==3.1.0
django-ninja==0.22.2
django-weasyprint==2.2.1
fonttools==4.45.1
html5lib==1.1
Pillow==10.1.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.13
pydantic-core==2.14.5
pydyf==0.8.0
pyphen==0.14.0
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.8.0
tzdata==2023.3
weasyprint==60.1
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
gunicorn==20.1.0
